LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-layout.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSInlineMediaControls</code> layout.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSInlineMediaControls({ width: 680, height: 50 });

const iconButtons = [
    mediaControls.playPauseButton,
    mediaControls.skipBackButton,
    mediaControls.airplayButton,
    mediaControls.pipButton,
    mediaControls.tracksButton,
    mediaControls.muteButton,
    mediaControls.fullscreenButton
];

// There is a 4-pixel difference in metrics between 1x and 2x due to the volume button having a different width.
const is1x = window.devicePixelRatio == 1;
const expectedTimeControlWidth = is1x ? 286 : 290;
const expectedVolumeSliderContainerX = is1x ? 428 : 432;

scheduler.frameDidFire = function()
{
    if (iconButtons.some(button => button.width == 0))
        return;

    shouldBeTrue("mediaControls.controlsBar.children[0] instanceof ButtonsContainer");
    shouldBeTrue("mediaControls.controlsBar.children[1] instanceof LayoutNode");
    shouldBeTrue("mediaControls.controlsBar.children[2] instanceof ButtonsContainer");

    shouldBe("mediaControls.timeControl.x", "118");

    shouldBeTrue("mediaControls.timeControl.width == expectedTimeControlWidth");

    shouldBe("mediaControls.controlsBar.children[0].children", "[mediaControls.playPauseButton, mediaControls.skipBackButton]");
    shouldBe("mediaControls.controlsBar.children[2].children", "[mediaControls.muteButton, mediaControls.airplayButton, mediaControls.pipButton, mediaControls.tracksButton, mediaControls.fullscreenButton]");

    shouldBeTrue("mediaControls.volumeSlider.parent.x == expectedVolumeSliderContainerX");

    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
